rename: 패키지명 변경
- 콘솔 애플리케이션 구현에 맞게 적절한 패키지명을 부여한다.
