+++ src/main/java/org/duckstudy/domain/lotto/LottoCount.java
@@ -1,4 +1,4 @@
-package org.duckstudy.model.lotto;
+package org.duckstudy.domain.lotto;
 
 import java.util.Objects;
