+++ src/test/java/org/duckstudy/domain/lotto/constant/WinningRankTest.java
@@ -1,4 +1,4 @@
-package org.duckstudy.model.lotto.constant;
+package org.duckstudy.domain.lotto.constant;
 
 import static org.assertj.core.api.Assertions.assertThat;
